django/django-ember-master/requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.9.7
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.3.3
gnureadline==6.3.3
ipdb==0.10.1
ipython==4.2.0
ipython-genutils==0.1.0
Markdown==2.6.6
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
simplegeneric==0.8.1
traitlets==4.2.1
